Renaming sharding tests to more inclusive language
https://developers.google.com/style/word-list?hl=en#dummy-variable
